final test

+++ Assets/Scripts/ScriptableObject/PlayerData.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerData", menuName = "Game/Player Data")]
 public class PlayerData : ScriptableObject
 {
-    public int maxHP;
+    public float maxHP;
     public float moveSpeed;
 }
